requirements.txt
Django==4.2.23
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow>=10.0.0
python-dotenv==1.0.1
django-cleanup==8.1.0
django-notifications-hq==1.8.3
django-allauth==0.61.1
django-multiselectfield 